backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
SQLAlchemy==2.0.35
psycopg[binary]==3.2.10
email-validator==2.2.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
alembic==1.14.0
faker==28.4.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pytest==8.3.3
httpx==0.27.2
requests>=2.31.0